requirements.txt
aniso8601==7.0.0
attrdict==2.0.1
attrs==19.3.0
base58==1.0.3
certifi==2019.9.11
chardet==3.0.4
cytoolz==0.10.0
Django==2.2.6
django-cors-headers==3.1.1
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.7.0
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
hexbytes==0.2.0
idna==2.8
importlib-metadata==0.23
ipfshttpclient==0.4.12
jsonschema==3.1.1
lru-dict==1.1.6
more-itertools==7.2.0
multiaddr==0.0.8
mysqlclient==1.4.4
netaddr==0.7.19
parsimonious==0.8.1
promise==2.2.1
protobuf==3.10.0
pycryptodome==3.9.0
pyrsistent==0.15.4
python-decouple==3.1
pytz==2019.3
requests==2.22.0
rlp==1.1.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
toolz==0.10.0
urllib3==1.25.6
varint==1.0.2
web3==5.2.1
websockets==7.0
zipp==0.6.0
